insert at index

+++ 0x13-more_singly_linked_lists/9-insert_nodeint.c
@@ -16,7 +16,7 @@ listint_t *insert_nodeint_at_index(listint_t **head, unsigned int idx, int n)
 
 	if (*head == NULL)
 		return (NULL);
-	if (idx < 0)
+	if ((int)idx < 0)
 		return (NULL);
 	while (tmp)
 	{
